requirements.txt
Django==2.1
django-toolbelt==0.0.1
djangorestframework==3.8.2
djangorestframework-msgpack==1.0.2
gunicorn==19.9.0
numpy==1.14.3
pandas==0.23.0
psycopg2==2.7.5
requests==2.18.4
scikit-learn==0.19.1
scikit-surprise==1.0.6
static3==0.7.0
